+++ lib/Dwoo/Compiler.php
@@ -1,16 +1,16 @@
 <?php
 /**
- * Copyright (c) 2013-2016
+ * Copyright (c) 2013-2017
  *
  * @category  Library
  * @package   Dwoo
  * @author    Jordi Boggiano <j.boggiano@seld.be>
  * @author    David Sanchez <david38sanchez@gmail.com>
  * @copyright 2008-2013 Jordi Boggiano
- * @copyright 2013-2016 David Sanchez
+ * @copyright 2013-2017 David Sanchez
  * @license   http://dwoo.org/LICENSE Modified BSD License
- * @version   1.3.0
- * @date      2016-09-23
+ * @version   1.3.2
+ * @date      2017-01-04
  * @link      http://dwoo.org/
  */
 
@@ -1391,7 +1391,7 @@ class Compiler implements ICompiler
         $substr = substr($in, $from, $to - $from);
 
         if ($this->debug) {
-            echo 'PARSE CALL : PARSING "<b>' . htmlentities(substr($in, $from, min($to - $from, 50))) . (($to - $from) > 50 ? '...' : '') . '</b>" @ ' . $from . ':' . $to . ' in ' . $curBlock . ' : pointer=' . $pointer . "\n";
+            echo 'PARSE CALL : PARSING "' . htmlentities(substr($in, $from, min($to - $from, 50))) . (($to - $from) > 50 ? '...' : '') . '" @ ' . $from . ':' . $to . ' in ' . $curBlock . ' : pointer=' . $pointer . "\n";
         }
         $parsed = '';
 
@@ -2321,16 +2321,25 @@ class Compiler implements ICompiler
      */
     protected function parseVar($in, $from, $to, $parsingParams = false, $curBlock = '', &$pointer = null)
     {
-        $methodCall = '';
-        $substr     = substr($in, $from, $to - $from);
+        $substr = substr($in, $from, $to - $from);
 
-        if (preg_match(
-            '#(\$?\.?[a-z0-9_:]*(?:(?:(?:\.|->)(?:[a-z0-9_:]+|(?R))|\[(?:[a-z0-9_:]+|(?R)|(["\'])[^\2]*?\2)\]))*)' . // var key
-            ($curBlock === 'root' || $curBlock === 'function' || $curBlock === 'namedparam' || $curBlock === 'condition' || $curBlock === 'variable' || $curBlock === 'expression' || $curBlock === 'delimited_string' ? '(\(.*)?' : '()') . // method call
-            ($curBlock === 'root' || $curBlock === 'function' || $curBlock === 'namedparam' || $curBlock === 'condition' || $curBlock === 'variable' || $curBlock === 'delimited_string' ? '((?:(?:[+/*%=-])(?:(?<!=)=?-?[$%][a-z0-9.[\]>_:-]+(?:\([^)]*\))?|(?<!=)=?-?[0-9.,]*|[+-]))*)' : '()') . // simple math expressions
-            ($curBlock !== 'modifier' ? '((?:\|(?:@?[a-z0-9_]+(?:(?::("|\').*?\5|:[^`]*))*))+)?' : '(())') . // modifiers
-            '#i', $substr, $match
-        )) {
+        // var key
+        $varRegex = '(\\$?\\.?[a-z0-9\\\\_:]*(?:(?:(?:\\.|->)(?:[a-z0-9\\\\_:]+|(?R))|\\[(?:[a-z0-9\\\\_:]+|(?R)|(["\'])[^\\2]*?\\2)\\]))*)';
+        // method call
+        $methodCall = ($curBlock === 'root' || $curBlock === 'function' || $curBlock === 'namedparam' || $curBlock === 'condition' || $curBlock === 'variable' || $curBlock === 'expression' || $curBlock === 'delimited_string' ? '(\(.*)?' : '()');
+        // simple math expressions
+        $simpleMathExpressions = ($curBlock === 'root' || $curBlock === 'function' || $curBlock === 'namedparam' || $curBlock === 'condition' || $curBlock === 'variable' || $curBlock === 'delimited_string' ? '((?:(?:[+\/*%=-])(?:(?<!=)=?-?[$%][a-z0-9\\\\.[\]>_:-]+(?:\([^)]*\))?|(?<!=)=?-?[0-9\.,]*|[+-]))*)' : '()');
+        // modifiers
+        $modifiers = $curBlock !== 'modifier' ? '((?:\|(?:@?[a-z0-9\\\\_]+(?:(?::("|\').*?\5|:[^`]*))*))+)?' : '(())';
+
+        $regex = '#';
+        $regex .= $varRegex;
+        $regex .= $methodCall;
+        $regex .= $simpleMathExpressions;
+        $regex .= $modifiers;
+        $regex .= '#i';
+
+        if (preg_match($regex, $substr, $match)) {
             $key = substr($match[1], 1);
 
             $matchedLength = strlen($match[0]);
@@ -2433,6 +2442,7 @@ class Compiler implements ICompiler
                 $output = $this->parseVarKey($key, $hasModifiers ? 'modifier' : $curBlock);
             }
 
+
             // methods
             if ($hasMethodCall) {
                 $ptr = 0;
@@ -2447,8 +2457,7 @@ class Compiler implements ICompiler
 
             if ($hasExpression) {
                 // expressions
-                preg_match_all('#(?:([+/*%=-])(=?-?[%$][a-z0-9.[\]>_:-]+(?:\([^)]*\))?|=?-?[0-9.,]+|\1))#i', $match[4], $expMatch);
-
+                preg_match_all('#(?:([+/*%=-])(=?-?[%$][a-z0-9\\\\.[\]>_:-]+(?:\([^)]*\))?|=?-?[0-9.,]+|\1))#i', $match[4], $expMatch);
                 foreach ($expMatch[1] as $k => $operator) {
                     if (substr($expMatch[2][$k], 0, 1) === '=') {
                         $assign = true;
@@ -2637,7 +2646,7 @@ class Compiler implements ICompiler
             } else {
                 $output = '(isset(' . $key . ')?' . $key . ':null)';
             }
-        } elseif (preg_match('#dwoo\.const\.([a-z0-9_:]+)#i', $key, $m)) {
+        } elseif (preg_match('#dwoo\\.const\\.([a-z0-9\\\\_:]+)#i', $key, $m)) {
             return $this->parseConstKey($m[1], $curBlock);
         } elseif ($this->scope !== null) {
             if (strstr($key, '.') === false && strstr($key, '[') === false && strstr($key, '->') === false) {