+++ lib/Dwoo/Template/File.php
@@ -1,16 +1,16 @@
 <?php
 /**
- * Copyright (c) 2013-2016
+ * Copyright (c) 2013-2017
  *
  * @category  Library
  * @package   Dwoo\Template
  * @author    Jordi Boggiano <j.boggiano@seld.be>
  * @author    David Sanchez <david38sanchez@gmail.com>
  * @copyright 2008-2013 Jordi Boggiano
- * @copyright 2013-2016 David Sanchez
+ * @copyright 2013-2017 David Sanchez
  * @license   http://dwoo.org/LICENSE Modified BSD License
- * @version   1.4.0
- * @date      2016-12-16
+ * @version   1.3.2
+ * @date      2017-01-03
  * @link      http://dwoo.org/
  */
 
@@ -41,7 +41,7 @@ class File extends Str
      *
      * @var array
      */
-    protected $includePath = null;
+    protected $includePath = array();
 
     /**
      * Resolved path cache when looking for a file in multiple include paths.
@@ -65,25 +65,13 @@ class File extends Str
      *                            template from others, if null it defaults to the filename+bits of the path
      * @param mixed  $includePath a string for a single path to look into for the given file, or an array of paths
      */
-    public function __construct($file, $cacheTime = null, $cacheId = null, $compileId = null, $includePath = null)
+    public function __construct($file, $cacheTime = null, $cacheId = null, $compileId = null, $includePath = array())
     {
-        $this->file      = $file;
-        $this->name      = basename($file);
-        $this->cacheTime = $cacheTime;
-
-        if ($compileId !== null) {
-            $this->compileId = str_replace('../', '__', strtr($compileId, '\\%?=!:;' . PATH_SEPARATOR, '/-------'));
-        }
-
-        if ($cacheId !== null) {
-            $this->cacheId = str_replace('../', '__', strtr($cacheId, '\\%?=!:;' . PATH_SEPARATOR, '/-------'));
-        }
-
-        if (is_string($includePath)) {
-            $this->includePath = array($includePath);
-        } elseif (is_array($includePath)) {
-            $this->includePath = $includePath;
-        }
+        parent::__construct($file, $cacheTime, $cacheId, $compileId);
+        $this->template = null;
+        $this->file     = $file;
+        $this->name     = basename($file);
+        $this->setIncludePath($includePath);
     }
 
     /**
@@ -154,15 +142,13 @@ class File extends Str
     {
         if ($this->resolvedPath !== null) {
             return $this->resolvedPath;
-        } elseif ($this->includePath === null) {
+        } elseif (array_filter($this->getIncludePath()) == array()) {
             return $this->file;
         } else {
-            foreach ($this->includePath as $path) {
+            foreach ($this->getIncludePath() as $path) {
                 $path = rtrim($path, DIRECTORY_SEPARATOR);
                 if (file_exists($path . DIRECTORY_SEPARATOR . $this->file) === true) {
-                    $this->resolvedPath = $path . DIRECTORY_SEPARATOR . $this->file;
-
-                    return $this->resolvedPath;
+                    return $this->resolvedPath = $path . DIRECTORY_SEPARATOR . $this->file;
                 }
             }