+++ tests/CoreTest.php
@@ -1,12 +1,4 @@
 <?php
-/**
- */
-
-/**
- */
-
-/**
- */
 
 namespace Dwoo\Tests
 {
@@ -199,6 +191,13 @@ namespace Dwoo\Tests
             ), $this->compiler));
         }
 
+        public function testGlobal()
+        {
+            $this->dwoo->addGlobal('test', 'value');
+            $tpl = new TemplateString('{$.test}');
+            $this->assertEquals('value', $this->dwoo->get($tpl, array(), $this->compiler));
+        }
+
         public function testSuperGlobals()
         {
             $_GET[5] = 'Yay';
@@ -385,6 +384,21 @@ namespace Dwoo\Tests
             $data = new TestCountableIterator(array('moo'));
             $this->assertEquals(true, $dwoo->isTraversable($data));
             $this->assertEquals(1, $dwoo->isTraversable($data, true));
+        }
+
+        public function testSetters()
+        {
+            $dwoo = new Core();
+            $dwoo->setCacheDir($this->cacheDir);
+            $dwoo->setCompileDir($this->compileDir);
+            $dwoo->setTemplateDir(__DIR__ . DIRECTORY_SEPARATOR . 'resources');
+
+            $this->assertThat($dwoo->get('extend1.html', array(), $this->compiler), new DwooConstraintStringEquals('foo
+child1
+toplevelContent1
+bar
+toplevelContent2
+baz'));
         }
     }
 }