+++ CHANGELOG.md
@@ -1,4 +1,22 @@
-## 1.3.1 (2016-XX-XX)
+## 1.3.2 (2017-01-05)
+`Added`
+* Add new tests: `CoreTest::testSetters`, `CoreTest::testGlobal`.
+* Add new methods: `Core::setTemplateDir` and `Core::getTemplateDir`.
+* Add new alias `js` for *javascript* available format in **escape** plugin.
+* Add new methods `Core::addGlobal` and `Core::getGlobals`.
+
+`Changed`
+* Properties `Core::$data` and `Core::$globals` are now protected.
+
+`Fixed`
+* Fix PHPUnit test: `CompilerTest::testConstants`.
+
+## 1.3.1 (2016-12-16)
+* Now fully compatible with PHP7.0 and PHP7.1.
+
+`Changed`
+* Rename class `Dwoo\Template\String` to `Dwoo\Template\Str`.
+
 `Fixed`
 * Fixing all adapters.
 * Fixing constant calls from classes with namespaces.